Summary:
Use a shift based method to extract 64-bits from a ymm register instead of pextrq. Shift instructions have better throughput and can execute on other ports.

Good performance improvement all around. For example, embedding dimension 32 sees a 16x improvement with prefetching on. Embedding dimension 28 was a slower case in production and it should see similar performance improvement. https://fburl.com/scuba/caffe2_operator_stats/jva9kutb

IACA analysis before: https://www.internalfb.com/intern/paste/P141371126

IACA analysis after:
https://www.internalfb.com/intern/paste/P142137812/

There is no change for avx512.

Results on BDW type 1 machine with turbo off. #thanks tylergraf for providing access to a broadwell test machine.

## Before Performance
```
bit_rate     2batch size    10  num rows         4000000   emb dim    32      avg length   100
64 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0608127 GB/s     effective b/w:         0.324334GB/s   time        0.0001772
            SLS        cache flushed    prefetch off     b/w 0.0510097 GB/s     effective b/w:         0.272052GB/s   time      0.000211254
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w  0.060629 GB/s     effective b/w:         0.323355GB/s   time      0.000177737
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0502243 GB/s     effective b/w:         0.267863GB/s   time      0.000214557
64 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0608545 GB/s     effective b/w:         0.324557GB/s   time      0.000177078
            SLS        cache flushed     prefetch on     b/w  0.056098 GB/s     effective b/w:         0.299189GB/s   time      0.000192092
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0599973 GB/s     effective b/w:         0.319985GB/s   time      0.000179608
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0559878 GB/s     effective b/w:         0.298602GB/s   time      0.000192471
32 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0609275 GB/s     effective b/w:         0.324947GB/s   time      0.000176866
            SLS        cache flushed    prefetch off     b/w 0.0505305 GB/s     effective b/w:         0.269496GB/s   time      0.000213257
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0607442 GB/s     effective b/w:         0.323969GB/s   time        0.0001774
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0502402 GB/s     effective b/w:         0.267948GB/s   time       0.00021449
32 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0604535 GB/s     effective b/w:         0.322418GB/s   time      0.000178253
            SLS        cache flushed     prefetch on     b/w 0.0562768 GB/s     effective b/w:         0.300143GB/s   time      0.000191482
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0600861 GB/s     effective b/w:         0.320459GB/s   time      0.000179343
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0558185 GB/s     effective b/w:         0.297699GB/s   time      0.000193054
bit_rate     2batch size    10  num rows         4000000   emb dim    64      avg length   100
64 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0513681 GB/s     effective b/w:         0.164378GB/s   time      0.000349633
            SLS        cache flushed    prefetch off     b/w 0.0471503 GB/s     effective b/w:         0.150881GB/s   time       0.00038091
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0516165 GB/s     effective b/w:         0.165173GB/s   time       0.00034795
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0462633 GB/s     effective b/w:         0.148042GB/s   time      0.000388213
64 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0512219 GB/s     effective b/w:          0.16391GB/s   time      0.000350631
            SLS        cache flushed     prefetch on     b/w 0.0495364 GB/s     effective b/w:         0.158516GB/s   time      0.000362562
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0513379 GB/s     effective b/w:         0.164281GB/s   time      0.000349839
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0488904 GB/s     effective b/w:         0.156449GB/s   time      0.000367352
32 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0514406 GB/s     effective b/w:          0.16461GB/s   time      0.000349141
            SLS        cache flushed    prefetch off     b/w 0.0472641 GB/s     effective b/w:         0.151245GB/s   time      0.000379992
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0516297 GB/s     effective b/w:         0.165215GB/s   time      0.000347862
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0468004 GB/s     effective b/w:         0.149761GB/s   time      0.000383758
32 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0513067 GB/s     effective b/w:         0.164181GB/s   time      0.000350052
            SLS        cache flushed     prefetch on     b/w 0.0493757 GB/s     effective b/w:         0.158002GB/s   time      0.000363742
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0514894 GB/s     effective b/w:         0.164766GB/s   time      0.000348809
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0491032 GB/s     effective b/w:          0.15713GB/s   time      0.000365761
bit_rate     2batch size    10  num rows         4000000   emb dim   128      avg length   100
64 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0462321 GB/s     effective b/w:        0.0821904GB/s   time      0.000699254
            SLS        cache flushed    prefetch off     b/w 0.0427932 GB/s     effective b/w:        0.0760767GB/s   time      0.000755448
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0465041 GB/s     effective b/w:        0.0826739GB/s   time      0.000695165
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0417967 GB/s     effective b/w:        0.0743052GB/s   time      0.000773459
64 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0463099 GB/s     effective b/w:        0.0823287GB/s   time       0.00069808
            SLS        cache flushed     prefetch on     b/w 0.0453137 GB/s     effective b/w:        0.0805577GB/s   time      0.000713427
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0461541 GB/s     effective b/w:        0.0820517GB/s   time      0.000700436
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0451812 GB/s     effective b/w:        0.0803222GB/s   time      0.000715518
32 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0463696 GB/s     effective b/w:        0.0824349GB/s   time      0.000697181
            SLS        cache flushed    prefetch off     b/w 0.0430231 GB/s     effective b/w:        0.0764856GB/s   time       0.00075141
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0464796 GB/s     effective b/w:        0.0826305GB/s   time       0.00069553
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0417706 GB/s     effective b/w:        0.0742589GB/s   time      0.000773941
32 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w  0.046273 GB/s     effective b/w:        0.0822631GB/s   time      0.000698636
            SLS        cache flushed     prefetch on     b/w 0.0452821 GB/s     effective b/w:        0.0805015GB/s   time      0.000713925
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0464904 GB/s     effective b/w:        0.0826495GB/s   time       0.00069537
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0454809 GB/s     effective b/w:        0.0808549GB/s   time      0.000710805
bit_rate     2batch size    10  num rows         4000000   emb dim   256      avg length   100
64 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0438867 GB/s     effective b/w:        0.0826103GB/s   time        0.0013914
            SLS        cache flushed    prefetch off     b/w 0.0415717 GB/s     effective b/w:        0.0782526GB/s   time       0.00146888
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0435573 GB/s     effective b/w:        0.0819902GB/s   time       0.00140192
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0412554 GB/s     effective b/w:        0.0776571GB/s   time       0.00148015
64 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0437105 GB/s     effective b/w:        0.0822786GB/s   time       0.00139701
            SLS        cache flushed     prefetch on     b/w 0.0432021 GB/s     effective b/w:        0.0813215GB/s   time       0.00141345
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0435649 GB/s     effective b/w:        0.0820045GB/s   time       0.00140168
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0430639 GB/s     effective b/w:        0.0810614GB/s   time       0.00141799
32 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0438375 GB/s     effective b/w:        0.0825176GB/s   time       0.00139296
            SLS        cache flushed    prefetch off     b/w 0.0411819 GB/s     effective b/w:        0.0775189GB/s   time       0.00148279
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0437995 GB/s     effective b/w:        0.0824462GB/s   time       0.00139417
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0412495 GB/s     effective b/w:        0.0776461GB/s   time       0.00148036
32 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0434089 GB/s     effective b/w:        0.0817108GB/s   time       0.00140672
            SLS        cache flushed     prefetch on     b/w 0.0432816 GB/s     effective b/w:        0.0814712GB/s   time       0.00141086
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0434341 GB/s     effective b/w:        0.0817582GB/s   time        0.0014059
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0430462 GB/s     effective b/w:        0.0810281GB/s   time       0.00141857
bit_rate     4batch size    10  num rows         4000000   emb dim    32      avg length   100
64 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w  0.103042 GB/s     effective b/w:         0.329734GB/s   time      0.000174298
            SLS        cache flushed    prefetch off     b/w 0.0822204 GB/s     effective b/w:         0.263105GB/s   time      0.000218437
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w  0.101497 GB/s     effective b/w:         0.324791GB/s   time      0.000176951
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0812517 GB/s     effective b/w:         0.260006GB/s   time      0.000221041
64 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w  0.101715 GB/s     effective b/w:         0.325489GB/s   time      0.000176571
            SLS        cache flushed     prefetch on     b/w  0.091995 GB/s     effective b/w:         0.294384GB/s   time      0.000195228
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w  0.102324 GB/s     effective b/w:         0.327437GB/s   time      0.000175521
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0940551 GB/s     effective b/w:         0.300976GB/s   time      0.000190952
32 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w  0.103053 GB/s     effective b/w:         0.329771GB/s   time      0.000174278
            SLS        cache flushed    prefetch off     b/w 0.0829468 GB/s     effective b/w:          0.26543GB/s   time      0.000216524
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w  0.101476 GB/s     effective b/w:         0.324724GB/s   time      0.000176987
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0822119 GB/s     effective b/w:         0.263078GB/s   time       0.00021846
32 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w   0.10304 GB/s     effective b/w:         0.329727GB/s   time      0.000174302
            SLS        cache flushed     prefetch on     b/w 0.0945706 GB/s     effective b/w:         0.302626GB/s   time      0.000189911
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w  0.101181 GB/s     effective b/w:         0.323779GB/s   time      0.000177504
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0935503 GB/s     effective b/w:         0.299361GB/s   time      0.000191982
bit_rate     4batch size    10  num rows         4000000   emb dim    64      avg length   100
64 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w  0.094158 GB/s     effective b/w:         0.167392GB/s   time      0.000343338
            SLS        cache flushed    prefetch off     b/w  0.075065 GB/s     effective b/w:         0.133449GB/s   time      0.000430667
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0938953 GB/s     effective b/w:         0.166925GB/s   time      0.000344298
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0761143 GB/s     effective b/w:         0.135314GB/s   time       0.00042473
64 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0933222 GB/s     effective b/w:         0.165906GB/s   time      0.000346413
            SLS        cache flushed     prefetch on     b/w 0.0892345 GB/s     effective b/w:         0.158639GB/s   time      0.000362281
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0937021 GB/s     effective b/w:         0.166582GB/s   time      0.000345008
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0895199 GB/s     effective b/w:         0.159146GB/s   time      0.000361127
32 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0942058 GB/s     effective b/w:         0.167477GB/s   time      0.000343164
            SLS        cache flushed    prefetch off     b/w 0.0771187 GB/s     effective b/w:           0.1371GB/s   time      0.000419198
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0937636 GB/s     effective b/w:         0.166691GB/s   time      0.000344782
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0764426 GB/s     effective b/w:         0.135898GB/s   time      0.000422905
32 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0945258 GB/s     effective b/w:         0.168046GB/s   time      0.000342002
            SLS        cache flushed     prefetch on     b/w 0.0900487 GB/s     effective b/w:         0.160087GB/s   time      0.000359006
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w   0.09298 GB/s     effective b/w:         0.165298GB/s   time      0.000347688
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0890978 GB/s     effective b/w:         0.158396GB/s   time      0.000362837
bit_rate     4batch size    10  num rows         4000000   emb dim   128      avg length   100
64 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0890019 GB/s     effective b/w:         0.167533GB/s   time      0.000686098
            SLS        cache flushed    prefetch off     b/w 0.0806084 GB/s     effective b/w:         0.151733GB/s   time      0.000757539
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0887028 GB/s     effective b/w:          0.16697GB/s   time      0.000688411
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0793714 GB/s     effective b/w:         0.149405GB/s   time      0.000769345
64 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w  0.088839 GB/s     effective b/w:         0.167226GB/s   time      0.000687356
            SLS        cache flushed     prefetch on     b/w 0.0861614 GB/s     effective b/w:         0.162186GB/s   time      0.000708717
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0885626 GB/s     effective b/w:         0.166706GB/s   time      0.000689501
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0854497 GB/s     effective b/w:         0.160847GB/s   time      0.000714619
32 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0891863 GB/s     effective b/w:          0.16788GB/s   time      0.000684679
            SLS        cache flushed    prefetch off     b/w 0.0806829 GB/s     effective b/w:         0.151874GB/s   time       0.00075684
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0888132 GB/s     effective b/w:         0.167178GB/s   time      0.000687556
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w 0.0793938 GB/s     effective b/w:         0.149447GB/s   time      0.000769128
32 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0883633 GB/s     effective b/w:         0.166331GB/s   time      0.000691056
            SLS        cache flushed     prefetch on     b/w  0.086265 GB/s     effective b/w:         0.162381GB/s   time      0.000707865
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w  0.088136 GB/s     effective b/w:         0.165903GB/s   time      0.000692838
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0861074 GB/s     effective b/w:         0.162085GB/s   time      0.000709161
bit_rate     4batch size    10  num rows         4000000   emb dim   256      avg length   100
64 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0861997 GB/s     effective b/w:         0.125381GB/s   time       0.00137513
            SLS        cache flushed    prefetch off     b/w 0.0811176 GB/s     effective b/w:         0.117989GB/s   time       0.00146129
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0860903 GB/s     effective b/w:         0.125222GB/s   time       0.00137688
  SLW(WEIGHTED)        cache flushed    prefetch off     b/w  0.080727 GB/s     effective b/w:         0.117421GB/s   time       0.00146836
64 bit indices with prefetching, lengths_sum 898
            SLS    cache not flushed     prefetch on     b/w 0.0861126 GB/s     effective b/w:         0.125255GB/s   time       0.00137652
            SLS        cache flushed     prefetch on     b/w 0.0846891 GB/s     effective b/w:         0.123184GB/s   time       0.00139966
  SLW(WEIGHTED)    cache not flushed     prefetch on     b/w 0.0861469 GB/s     effective b/w:         0.125305GB/s   time       0.00137597
  SLW(WEIGHTED)        cache flushed     prefetch on     b/w 0.0846334 GB/s     effective b/w:         0.123103GB/s   time       0.00140058
32 bit indices, lengths_sum 898
            SLS    cache not flushed    prefetch off     b/w 0.0863298 GB/s     effective b/w:         0.125571GB/s   time       0.00137306
            SLS        cache flushed    prefetch off     b/w 0.0811525 GB/s     effective b/w:          0.11804GB/s   time       0.00146066
  SLW(WEIGHTED)    cache not flushed    prefetch off     b/w 0.0859662 GB/s     effective b/w:         0.125042GB/s   time       0.00137887
…